split asset exchange test and cache cdn
